+++ addOns/client/CHANGELOG.md
@@ -4,7 +4,8 @@ All notable changes to this add-on will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## Unreleased
-
+### Fixed
+- Prevent temporary GUI hang when stopping the Client Spider.
 
 ## [0.27.0] - 2026-06-12
 ### Added